requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.7.3
idna==2.10
multidict==4.7.6
psycopg2
yarl==1.5.1
